Origin

+++ fruitking/WebContent/tab/view.jsp
@@ -10,12 +10,6 @@
 <meta charset="UTF-8">
 <title>notify</title>
 <link rel="stylesheet" href="../css/bootstrap.css">
-<style type="text/css">
-	#content {
-		min-height: 300px;
-		text-align: left;
-	}
-</style>
 </head>
 <body>
 	<%
@@ -31,9 +25,11 @@
 			script.println("</script>");
 		}
 		NotifyDTO notify = new NotifyDAO().getNotifyDTO(id);
+		NotifyDTO nextNotify = new NotifyDAO().getNextNotify(id);
+		NotifyDTO priorNotify = new NotifyDAO().getPriorNotify(id);
 	%>
-	<div class="container">
-		<div class="row">
+	<div>
+		<div style="background-color: rgb(253,255,244);">
 			<table class="table table-striped" style="text-align: center;	border: 1px solid #dddddd;">
 				<thead>
 					<tr>
@@ -51,27 +47,45 @@
 					</tr>
 					<tr>
 						<td>내용</td>
-						<td id="content" colspan="2" style="min-height: 300px; text-align: left;"><%= notify.getContent() %></td>
+						<td colspan="2"><div style="min-height: 200px; text-align: left;"><%= notify.getContent() %></div></td>
 					</tr>
-					<tr>
-						<td>윗 글</td>
-						<td colspan="2""><a href="view.jsp?id=<%=id + 1 %>">
-						<% NotifyDTO notifyNext = new NotifyDAO().getNotifyDTO(id + 1); %>
-						<%= notifyNext.getTitle() %>
-						</a></td>
-					</tr>
-					<tr>
-						<td>아랫 글</td>
-						<td colspan="2""><a href="view.jsp?id=<%=id - 1 %>">
-						<% NotifyDTO notifyPrior = new NotifyDAO().getNotifyDTO(id - 1); %>
-						<%= notifyPrior.getTitle() %>
-						</a></td>
-					</tr>
+					<%
+						if(nextNotify.getId() == notify.getId()){
+					%>
+						<tr>
+							<td>윗글</td>
+							<td colspan="2">윗글이 없습니다.</td>
+						</tr>
+					<%
+						} else {
+					%>
+						<tr>
+							<td>윗글</td>
+							<td colspan="2"><a href="view.jsp?id=<%=nextNotify.getId() %>"><%= nextNotify.getTitle() %></a></td>
+						</tr>
+					<%
+						}
+						if(priorNotify.getId() == notify.getId()){
+					%>
+						<tr>
+							<td>아랫글</td>
+							<td colspan="2">아랫글이 없습니다.</td>
+						</tr>
+					<%
+						} else {
+					%>
+						<tr>
+							<td>아랫글</td>
+							<td colspan="2"><a href="view.jsp?id=<%=priorNotify.getId()%>"><%= priorNotify.getTitle() %></a></td>
+						</tr>
+					<%
+						}
+					%>
 				</tbody>
 			</table>
 			<a href="notify.jsp" class="btn btn-primary">목록</a>
-			<a href="update.jsp?id=<%=id %>" class="btn btn-primary">수정</a>
-			<a href="deleteAction.jsp?id=<%=id %>" class="btn btn-primary">삭제</a>
+			<a href="updateNotify.jsp?id=<%=id %>" class="btn btn-primary">수정</a>
+			<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?id=<%=id %>" class="btn btn-primary pull-right">삭제</a>
 		</div>
 	</div>
